Crate concourse_resource

Helper to implement a Concourse resource in Rust

Concourse documentation

Modules

internal
Internal types used to wrap inputs and outputs. They should not be built directly but are used by macros

Macros

create_resource
Macro that will build the main function from a struct implementing the Resource trait

Structs

BuildMetadata
When used in a "get" or "put" step, metadata about the running build is made available via environment variables.

Empty
Empty value that can be used as InParams, InMetadata, OutParams or OutMetadata for a Resource

InOutput
Output of the "in" step of the resource

OutOutput
Output of the "out" step of the resource

Traits

IntoMetadataKV
Trait for Metadata to be usable as Concourse Metadata. This trait can be derived if the base struct implement serde::Deserialize

Resource
The methods and associated types needed to implement a resource

Derive Macros

IntoMetadataKV